+++ contracts/core/interfaces/ISetFactory.sol
@@ -48,7 +48,7 @@ interface ISetFactory {
     function create(
         address[] _components,
         uint[] _units,
-        uint _naturalUnit,
+        uint256 _naturalUnit,
         string _name,
         string _symbol
     )
